use package info_field_ids

+++ lib/SPVM/Builder/include/spvm_my.h
@@ -7,11 +7,11 @@ struct SPVM_my {
   SPVM_OP* op_my;
   SPVM_TYPE* type;
   SPVM_OP* op_name;
-  uint16_t index;
-  uint16_t var_id;
-  uint8_t runtime_type;
-  uint8_t type_width;
-  uint8_t is_tmp;
+  int32_t index;
+  int32_t var_id;
+  int32_t runtime_type;
+  int32_t type_width;
+  int8_t is_tmp;
 };
 
 SPVM_MY* SPVM_MY_new();
